Rename train_task to training_task
